Cache JAXB contexts (FLIB-36)
We need to ensure we don't create new JAXB contexts every time we parse an IDOL response as this is very expensive.
This is a cheap fix using a ConcurrentHashMap, which satisfies the requirements for the upcoming FIND 11.2 release.
The new code was not added to a separate class because there did not seem sufficient immediate benefit in doing so.
A more comprehensive solution is forthcoming.

[rev. matthew.gordon]